+++ changelog.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is loosely based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+- **Added** [#57](https://github.com/codyburleson/cburleson-pwa/issues/57) Page: are-mashups-the-new-portals-and-gadgets-the-new-portlets
 ## [1.1.3] - 2019-12-26
 
 - **Added** [#66](https://github.com/codyburleson/cburleson-pwa/issues/66) Pages: memories-of-momma-1 through 6